Shuffle around init code to bootstrap transpiled code in IronPython
